engines: Add a customer-socket engine for Help.com
Artillery's basic implementation of socket.io was not fitting with Help.com,
mainly it's restriction that in order to receive messages, you must first emit
a message on the socket.  Since Help.com's websockets are unidirectional,
this did not work.  Plus, lots of cruft was removed since Help.com does not
use namespaces or channels.  This new implementation allows for listeners
to be set up, THEN it will execute any HTTP actions that generate websocket
messages.